Enable preprocessor
