+++ qna/src/app/domain/__init__.py
@@ -4,4 +4,4 @@ from .universalencoder import UniversalEncoder
 from .keywordmatcher import KeywordMatcher
 from .questionmatcher import AbstractQuestionMatcher
 from .summarisation import AbstractSummarisation
-from.t5 import T5
+from .t5 import T5
